requirements.txt
# Core dependencies
python-dotenv==1.0.0
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
fpdf2>=2.7.0
langchain>=0.0.350
langchain-community>=0.0.10
transformers>=4.35.0
torch>=2.0.0
tavily-python>=0.3.0
pydantic>=2.5.0,<3.0.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6
streamlit-ace>=0.1.1
streamlit-components-template>=0.0.1
psutil>=5.9.0
watchdog>=3.0.0
cryptography>=41.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
ratelimit>=2.2.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
# Global Database Integration Dependencies
xmltodict>=0.13.0
lxml>=4.9.0
biopython>=1.81
requests-cache>=1.1.0
retrying>=1.3.4
aiohttp>=3.8.0
aiofiles>=23.0.0
# Enhanced Data Processing
pyarrow>=13.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
# Network and API enhancements
httpx>=0.25.0
urllib3>=2.0.0
certifi>=2023.7.28
typing-extensions==4.8.0
pillow==10.1.0
